use accept_data for setting position_valid
Check the return value from accept_data to determine if we want to
use the decoded position.
Avoid a position from local CPR with a synthetic MLAT message being used
when the position may still be valid and either odd or even CPR already
stale allowing the MLAT CPR to be used.
Set the position expiration correctly for global surface CPR, the
previous code could have an up to 50 second old CPR being used, which
would result in combine_validity setting an expiry time for the position
only 10 seconds in the future.