Rearrange config upgrade so upgrading a v6 config file disables adaptive gain.
This is needed for the case where a v5 package with an unmodified
config file is upgraded. dpkg will install the v6 default config
file automatically so it will already be a v6 config by the time
that upgrade-config runs, but we do still need to go in and
disable adaptive gain.